Split vcffile out and enabled opening it for writing with a bgzip option.
